# backend/fastapi/openapi-fastapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workma!te AI API",
    "description": "회의체 운영 AI Agent 서비스의 **AI · 실시간 · 지식그래프** 백엔드 (FastAPI).\n\n- **인증**: Spring이 발급한 JWT access token을 `Authorize`(HTTPBearer)에 입력하세요. (`JWT_SECRET`을 Spring과 공유, HS256)\n- **스트리밍**: 다수 엔드포인트가 SSE(`text/event-stream`)로 토큰 단위 응답합니다.\n- **데이터 스코프**: AI 도구는 사용자의 소속 회의체로 접근을 제한합니다(IDOR 이중 방어).\n- 인증·도메인 CRUD는 별도 Spring Boot 서버(`/api/v1`)가 담당합니다.",
    "contact": {
      "name": "Team No.9"
    },
    "license": {
      "name": "Internal"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sync/retry": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "실패 동기화 재시도 트리거",
        "description": "agent_logs의 실패 항목을 재시도합니다.\n- max_retries: 이 횟수 미만인 항목만 대상",
        "operationId": "trigger_retry_api_sync_retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "max_retries",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3,
              "title": "Max Retries"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/user/{user_id}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "User 동기화 (내부 전용)",
        "description": "특정 User를 Neo4j에 동기화합니다. Spring Boot에서 유저 생성/수정 시 호출합니다.\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "sync_user_manual_api_sync_user__user_id__post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sync"
        ],
        "summary": "User 삭제 동기화 (내부 전용)",
        "description": "User 노드 및 연결된 모든 관계를 Neo4j에서 삭제합니다.\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "delete_user_sync_api_sync_user__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/department": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Department 동기화 (내부 전용)",
        "description": "Department 노드를 Neo4j에 수동으로 동기화합니다.\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "sync_department_manual_api_sync_department_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/company": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Company 동기화 (내부 전용)",
        "description": "Company 노드를 Neo4j에 수동으로 동기화합니다.\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "sync_company_manual_api_sync_company_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/meeting/{meeting_id}/delete": {
      "delete": {
        "tags": [
          "sync"
        ],
        "summary": "Meeting 삭제 동기화 (내부 전용)",
        "description": "Meeting 노드 및 연결된 모든 관계를 Neo4j에서 삭제합니다.\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "delete_meeting_sync_api_sync_meeting__meeting_id__delete_delete",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/meeting/{meeting_id}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Meeting 동기화 (내부 전용)",
        "description": "특정 Meeting을 Neo4j에 수동으로 동기화합니다.\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "sync_meeting_manual_api_sync_meeting__meeting_id__post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/session/{session_id}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Session 동기화 (내부 전용)",
        "description": "특정 Session을 Neo4j에 수동으로 동기화합니다(참석자·논의 안건 포함).\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "sync_session_manual_api_sync_session__session_id__post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sync"
        ],
        "summary": "Session 삭제 동기화 (내부 전용)",
        "description": "Session 삭제를 Neo4j에 전파합니다 (DATA-4).\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "delete_session_sync_api_sync_session__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/all": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "전체 동기화 (부트스트랩·복구)",
        "description": "PostgreSQL의 모든 엔티티를 Neo4j에 전체 동기화합니다.\nNeo4j 재설치 또는 데이터 불일치 복구 시 사용합니다.\n백그라운드에서 실행되며 즉시 응답을 반환합니다.",
        "operationId": "sync_all_api_sync_all_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sync/agenda/{agenda_id}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Agenda 동기화 (내부 전용)",
        "description": "특정 Agenda를 Neo4j에 수동으로 동기화합니다.\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "sync_agenda_manual_api_sync_agenda__agenda_id__post",
        "parameters": [
          {
            "name": "agenda_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agenda Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sync"
        ],
        "summary": "Agenda 삭제 동기화 (내부 전용)",
        "description": "Agenda 삭제를 Neo4j에 전파합니다 (DATA-4).\n\n내부 전용 — X-Internal-Secret 헤더 검증(verify_internal) 필요.",
        "operationId": "delete_agenda_sync_api_sync_agenda__agenda_id__delete",
        "parameters": [
          {
            "name": "agenda_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agenda Id"
            }
          },
          {
            "name": "x-internal-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/minutes/{minutes_id}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Minutes 수동 동기화",
        "description": "특정 Minutes를 Neo4j에 수동으로 동기화합니다. 인증 필요.",
        "operationId": "sync_minutes_manual_api_sync_minutes__minutes_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "minutes_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Minutes Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sync/report/{report_id}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Report 수동 동기화",
        "description": "특정 Report를 Neo4j에 수동으로 동기화합니다. 인증 필요.",
        "operationId": "sync_report_manual_api_sync_report__report_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}